Move Tower defense exapmle to separate example project
