+++ AKE/Main/include/Main/Example2D.h
@@ -9,6 +9,4 @@ namespace Example2D {
   void RotatedTriangles(std::size_t i_window_width, std::size_t i_window_height, std::size_t i_triangles_count);
 
   void CollisionDetection(std::size_t i_window_width, std::size_t i_window_height, std::size_t i_objects_count);
-
-  void TowerDefenseGame(std::size_t i_window_width, std::size_t i_window_height);
 }
